ref: Move warning to url-encoding, add tests
